# spec/swagger.yaml
definitions:
...
  acpagent.RuntimeStatus:
    properties:
      acp_session_id:
        type: string
      agent_id:
        type: string
      models:
        $ref: '#/definitions/acpclient.ModelState'
      project_path:
        type: string
      session_id:
        type: string
      state:
        type: string
    type: object
  acpclient.ModelInfo:
    properties:
      description:
        type: string
      id:
        type: string
      name:
        type: string
    type: object
  acpclient.ModelState:
    properties:
      available_models:
        items:
          $ref: '#/definitions/acpclient.ModelInfo'
        type: array
      current_model_id:
        type: string
      supported:
        type: boolean
    type: object
...
  handlers.ErrorResponse:
    properties:
      message:
        type: string
    type: object
...
  handlers.acpRuntimeModelRequest:
    properties:
      model_id:
        type: string
    type: object
...
info:
  contact: {}
  title: Memoh API
  version: 1.0.0
paths:
...
  /bots/{bot_id}/sessions/{session_id}/acp-runtime:
    get:
      parameters:
      - description: Bot ID
        in: path
        name: bot_id
        required: true
        type: string
      - description: Session ID
        in: path
        name: session_id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/acpagent.RuntimeStatus'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Get ACP session runtime state
      tags:
      - acp
    post:
      parameters:
      - description: Bot ID
        in: path
        name: bot_id
        required: true
        type: string
      - description: Session ID
        in: path
        name: session_id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/acpagent.RuntimeStatus'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Ensure ACP session runtime is started
      tags:
      - acp
  /bots/{bot_id}/sessions/{session_id}/acp-runtime/model:
    patch:
      parameters:
      - description: Bot ID
        in: path
        name: bot_id
        required: true
        type: string
      - description: Session ID
        in: path
        name: session_id
        required: true
        type: string
      - description: ACP model selection
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handlers.acpRuntimeModelRequest'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/acpagent.RuntimeStatus'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Set ACP session runtime model
      tags:
      - acp
...
swagger: "2.0"